// src/hass-types/src/data/selector.d.ts
export interface SelectOption {
    value: any;
    label: string;
    disabled?: boolean;
}
export interface SelectSelector {
    select: {
        multiple?: boolean;
        custom_value?: boolean;
        mode?: "list" | "dropdown";
        options: readonly string[] | readonly SelectOption[];
        translation_key?: string;
        sort?: boolean;
        reorder?: boolean;
    } | null;
}